+++ CHANGES.md
@@ -17,6 +17,10 @@
 -   Export a few convenience types: `IAnySubFormAccessor`,
     `IAnyRepeatingFormAccessor`, `IAnyRepeatingFormIndexedAccessor`.
 
+-   Add a `maxZeroesPadding` option to the `decimal` renderer.
+    `maxZeroesPadding` can be set to limit the number of trailing zeros. This to
+    display a flexible minimum amount of decimals. Show more if we have more.
+
 # 1.25.0
 
 -   You can pass the optional parameter `liveOnly` to the `processAll` function